adding mockNimEnabledProject mocking function
